The DFC Rebirth
Added Fancy GUIs to all DFC Machines
Added Dimension Nuke water drain/fill config
Added Silex Crystal JEI
Added Astatine Ingot
Added Stabilizer Stacking
Added Desh Stamps
Added Fluid Pipe Click-Feature
Added New Stabelizer Lenses

Changed Cryogel Texture
Changed Sellaline no longer sets you on fire
Changed Order of Fluids
Changed Order of Items
Changed DFC Rendering
Changed DFC Recipes

Fixed DFC Emitter GUI
Fixed DFC Catalysts not working
Fixed Schists generation - idea by Drillgon
Fixed dublicate infinite batteries
Fixed Energy Control Crash
Fixed Lignite Fluorite mixup
Fixed Jungle Dungeon spawning
Fixed Smore Crafting
Fixed Orbus output

Removed Unused Plasma Heater Slots
Removed Unused Textures/Models

+++ src/main/java/com/hbm/tileentity/machine/TileEntityCrateDesh.java
@@ -104,7 +104,7 @@ public class TileEntityCrateDesh extends TileEntityLockableBase implements ITick
 			
 			ItemStack result = FurnaceRecipes.instance().getSmeltingResult(inventory.getStackInSlot(i));
 			
-			int requiredEnergy = DFCRecipes.getRequiredFlux(inventory.getStackInSlot(i));
+			long requiredEnergy = DFCRecipes.getRequiredFlux(inventory.getStackInSlot(i));
 			if(requiredEnergy > -1 && energy > requiredEnergy){
 				result = DFCRecipes.getOutput(inventory.getStackInSlot(i));
 			}
